Create file
This project gives the inference about what the consumer demands from the shop and what the manufacturers of the spices can do to increase their sales by attracting the customers. The dataset taken for the project is a public survey of people buying the branded spices like ITC, Sakthi, Aachi etc., at a mall or super markets. The dataset has the feedback of the customers regarding the features of the products such as brand, price, advertisements etc. The descriptive analysis is done to understand the buying patterns and behaviors. The diagnostic analysis is done to understand what are the factors are influencing the consumers and why it is also inferred.